+++ ASP/CRM microservices architecture/crm-microservices/OrderService/OrderService/GrpcServices/OrderGrpcServices.cs
@@ -5,6 +5,8 @@ using CRMSolution.Grpc.Tasks;
 using CRMSolution.Grpc.Users;
 using Grpc.Core;
 using OrderService.Data.Models;
+using OrderService.Data.Repository.OrderResp;
+using OrderService.Data.Validators.Order;
 using OrderService.DTO.Requests; // <-- Только этот using правильный!
 using OrderService.Services.Interfaces;
 using OrderService.DTO.Requests.Order;
@@ -20,10 +22,11 @@ public class OrderGrpcService : CRMSolution.Grpc.Orders.OrderGrpcService.OrderGr
     private readonly ClientGrpcService.ClientGrpcServiceClient _clientGrpcClient;
     private readonly TaskGrpcService.TaskGrpcServiceClient _taskGrpcService;
     private readonly IMapper _mapper;
+    private readonly IOrderRep _orderRep;
 
     public OrderGrpcService(IOrderService orderService,  ILogger<OrderGrpcService> logger, UserService.UserServiceClient userGrpcClient,
         ClientGrpcService.ClientGrpcServiceClient clientGrpcClient,  TaskGrpcService.TaskGrpcServiceClient taskGrpcService,
-        IMapper mapper)
+        IMapper mapper,  IOrderRep orderRep)
     {
         _orderService = orderService;
         _logger = logger;
@@ -31,10 +34,19 @@ public class OrderGrpcService : CRMSolution.Grpc.Orders.OrderGrpcService.OrderGr
         _clientGrpcClient = clientGrpcClient;
         _taskGrpcService = taskGrpcService;
         _mapper = mapper;
+        _orderRep = orderRep;
     }
 
     public override async Task<OrderDto> GetOrderById(GetOrderByIdRequest request, ServerCallContext context)
     {
+        var validator = new FindOrderValidator(_orderRep);
+        var result = validator.Validate(request);
+
+        if (!result.IsValid)
+        {
+            var errorMessages = string.Join(" | ", result.Errors.Select(e => e.ErrorMessage));
+            throw new RpcException(new Status(StatusCode.InvalidArgument, errorMessages));
+        }
         var order = await _orderService.GetByIdAsync(request.OrderId);
 
         return new OrderDto
@@ -50,17 +62,41 @@ public class OrderGrpcService : CRMSolution.Grpc.Orders.OrderGrpcService.OrderGr
     public override async Task<CreateOrderResponse> CreateOrder(CreateOrderRequest request,
         ServerCallContext context)
     {
+        var validator = new CreateOrderValidator();
+        var result = validator.Validate(request);
+
+        if (!result.IsValid)
+        {
+            var errorMessages = string.Join(" | ", result.Errors.Select(e => e.ErrorMessage));
+            throw new RpcException(new Status(StatusCode.InvalidArgument, errorMessages));
+        }
         return await _orderService.CreateOrder(request);
     }
 
     public override async Task<ChangeOrderDataResponse> ChangeOrderData(ChangeOrderDataRequest request,
         ServerCallContext context)
     {
+        var validator = new ChangeOrderDataValidator(_orderRep);
+        var result = validator.Validate(request);
+
+        if (!result.IsValid)
+        {
+            var errorMessages = string.Join(" | ", result.Errors.Select(e => e.ErrorMessage));
+            throw new RpcException(new Status(StatusCode.InvalidArgument, errorMessages));
+        }
         return await _orderService.ChangeDataOrder(request);
     }
 
     public override async Task<DeleteOrderResponse> DeleteOrder(DeleteOrderRequest request, ServerCallContext context)
     {
+        var validator = new DeleteOrderValidator(_orderRep);
+        var result = validator.Validate(request);
+
+        if (!result.IsValid)
+        {
+            var errorMessages = string.Join(" | ", result.Errors.Select(e => e.ErrorMessage));
+            throw new RpcException(new Status(StatusCode.InvalidArgument, errorMessages));
+        }
         await _orderService.DeleteOrder(request);
         
         return new DeleteOrderResponse
