+++ ASP/CRM microservices architecture/crm-microservices/TaskService/TaskService/GrpcServices/TaskGrpcServices.cs
@@ -1,6 +1,9 @@
-﻿using Grpc.Core;
+﻿using AutoMapper;
+using CRMSolution.Data.Validators.Tasks;
+using Grpc.Core;
 using CRMSolution.Grpc.Tasks;
-using TaskService.Data.Models; // <-- Правильный namespace
+using TaskService.Data.Models;
+using TaskService.Data.Repository.TasksRep; // <-- Правильный namespace
 using TaskService.Services.Interfaces;
 
 namespace TaskService.GrpcServices;
@@ -8,10 +11,12 @@ namespace TaskService.GrpcServices;
 public class TaskGrpcService : CRMSolution.Grpc.Tasks.TaskGrpcService.TaskGrpcServiceBase
 {
     private readonly ITasksService _tasksService;
+    private readonly ITasksRep _tasksRep;
 
-    public TaskGrpcService(ITasksService tasksService)
+    public TaskGrpcService(ITasksService tasksService, ITasksRep tasksRep)
     {
         _tasksService = tasksService;
+        _tasksRep = tasksRep;
     }
 
     public override async Task<GetTasksByUserIdsResponse> GetTasksByUserIds(GetTasksByUserIdsRequest request,
@@ -27,6 +32,14 @@ public class TaskGrpcService : CRMSolution.Grpc.Tasks.TaskGrpcService.TaskGrpcSe
     }
     public override async Task<GetTaskByIdResponse> GetTaskById(GetTaskByIdRequest request, ServerCallContext context)
     {
+        var validator = new FindTaskValidator(_tasksRep);
+        var result = validator.Validate(request);
+
+        if (!result.IsValid)
+        {
+            var errorMessages = string.Join(" | ", result.Errors.Select(e => e.ErrorMessage));
+            throw new RpcException(new Status(StatusCode.InvalidArgument, errorMessages));
+        }
         var task = await _tasksService.GetByIdAsync(request.Id);
 
         return new GetTaskByIdResponse
@@ -40,6 +53,14 @@ public class TaskGrpcService : CRMSolution.Grpc.Tasks.TaskGrpcService.TaskGrpcSe
     
     public override async Task<CreateTaskResponse> CreateTask(CreateTaskRequest request, ServerCallContext context)
     {
+        var validator = new CreateTaskValidator();
+        var result = validator.Validate(request);
+
+        if (!result.IsValid)
+        {
+            var errorMessages = string.Join(" | ", result.Errors.Select(e => e.ErrorMessage));
+            throw new RpcException(new Status(StatusCode.InvalidArgument, errorMessages));
+        }
         return await _tasksService.CreateTaskAsync(request.OrderId, request.Description, request.DueDate.ToDateTime(), request.Title);
     }
 
@@ -52,12 +73,28 @@ public class TaskGrpcService : CRMSolution.Grpc.Tasks.TaskGrpcService.TaskGrpcSe
     public override async Task<TaskInfo> UpdateTask(UpdateTaskRequest request,
         ServerCallContext context)
     {
+        var validator = new UpdateTaskValidator(_tasksRep);
+        var result = await validator.ValidateAsync(request);
+
+        if (!result.IsValid)
+        {
+            var errorMessages = string.Join(" | ", result.Errors.Select(e => e.ErrorMessage));
+            throw new RpcException(new Status(StatusCode.InvalidArgument, errorMessages));
+        }
         return await _tasksService.UpdateTaskAsync(request);
     }
     
     public override async Task<DeleteTaskResponse> DeleteTask(DeleteTaskRequest DeleteTaskRequest,
         ServerCallContext context)
     {
+        var validator = new DeleteTaskValidator(_tasksRep);
+        var result = validator.Validate(DeleteTaskRequest);
+
+        if (!result.IsValid)
+        {
+            var errorMessages = string.Join(" | ", result.Errors.Select(e => e.ErrorMessage));
+            throw new RpcException(new Status(StatusCode.InvalidArgument, errorMessages));
+        }
         return await _tasksService.DeleteTaskAsync(DeleteTaskRequest);
     }
